pre-commit: drop no-commit-to-branch
By design the no-commit-to-branch hook triggers a failure when
pre-commit is run from the 'main' branch. This works fine locally, but
makes it impossible to run CI against main, which is something we want
to keep doing. Let's drop the hook and rely on branch protection to
prevent pushes to the wrong branch.