cii: allow for skipping non-lts releases on some platform
Releases can have a 'release-lts' flag that makes it easy to exclude
them using an 'exclude' stanza [1].

Also skip Jammy on Oracle for the moment as images are not available yet.
A better way would be checking for the availability of an image on a
given platform and mark the job as "unstable" (yellow) if there's no
image. This is however out of the scope of this change.

[1] https://docs.openstack.org/infra/jenkins-job-builder/definition.html#project