metrics: retry the entire shell builder on failure
This introduces a new retry mechanism I plan to use on Jenkins jobs for
which there isn't really value in the individual failures. This is the
case of e.g. hourly jobs collecting metrics data over the network.

Bump the timeout for the metrics jobs to 60 minutes, to take the
retrials into account.